# docs/swagger.yaml
basePath: /api
definitions:
  models.BankTransactionRequest:
    properties:
      account_number:
        type: string
      amount:
        type: integer
      bank_code:
        type: string
      description:
        type: string
    type: object
  models.TransactionAmount:
    properties:
      amount:
        type: integer
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a sample top-up system server.
  title: Top Up System API
  version: "1.0"
paths:
  /transactions/topup/bank/{va}:
    post:
      consumes:
      - application/json
      description: Top up wallet via bank transfer
      parameters:
      - description: Virtual Account Number
        in: path
        name: va
        required: true
        type: integer
      - description: Bank transaction request
        in: body
        name: bank
        required: true
        schema:
          $ref: '#/definitions/models.BankTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Top up wallet via bank transfer
      tags:
      - transactions
  /transactions/topup/direct/{id}:
    post:
      consumes:
      - application/json
      description: Top up wallet directly
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Amount to top up
        in: body
        name: amount
        required: true
        schema:
          $ref: '#/definitions/models.TransactionAmount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Top up wallet directly
      tags:
      - transactions
  /transactions/wallet/{id}:
    get:
      description: Get transactions by wallet ID
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get transactions by wallet ID
      tags:
      - transactions
  /users:
    get:
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all users
      tags:
      - users
  /users/{id}:
    get:
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get user by ID
      tags:
      - users
  /wallets:
    get:
      description: Get all wallets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all wallets
      tags:
      - wallets
  /wallets/{id}:
    get:
      description: Get wallet by ID
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get wallet by ID
      tags:
      - wallets
  /wallets/user/{id}:
    get:
      description: Get wallet by User ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get wallet by User ID
      tags:
      - wallets
  /wallets/va/{id}:
    get:
      description: Get virtual account by Wallet ID
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get virtual account by Wallet ID
      tags:
      - wallets
swagger: "2.0"